Mudado tipo do policingTime para int

+++ projeto/src/patrol_allocation/Distribution.java
@@ -83,7 +83,7 @@ public class Distribution {
      *
      * @return The policing time.
      */
-    public double policingTime() {
+    public int policingTime() {
         return policingTime;
     }
